+++ src/main/java/domain/piece/Po.java
@@ -5,7 +5,6 @@ import static domain.board.Direction.LEFT;
 import static domain.board.Direction.RIGHT;
 import static domain.board.Direction.UP;
 
-import domain.PieceType;
 import domain.board.Board;
 import domain.board.Direction;
 import domain.board.Node;
